+++ app/cells/folio/devise/invitations/show_cell.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Folio::Devise::Invitations::ShowCell < Folio::Devise::ApplicationCell
+  def show
+    render if model.present?
+  end
+end