+++ app/controllers/folio/users/invitations_controller.rb
@@ -3,11 +3,16 @@
 class Folio::Users::InvitationsController < Devise::InvitationsController
   include Folio::Users::DeviseControllerBase
 
-  def new
-    fail ActionController::MethodNotAllowed, ""
+  def show
+    if session[:folio_user_invited_email]
+      @email = session[:folio_user_invited_email]
+    else
+      redirect_to new_user_invitation_path
+    end
   end
 
-  def create
-    fail ActionController::MethodNotAllowed, ""
+  def after_invite_path_for(_inviter, resource)
+    session[:folio_user_invited_email] = resource.email
+    user_invitation_path
   end
 end