+++ lib/folio/engine.rb
@@ -46,7 +46,7 @@ module Folio
     config.folio_users = false
     config.folio_users_require_phone = false
     config.folio_users_confirmable = false
-    config.folio_users_registerable = true
+    config.folio_users_publicly_invitable = true
     config.folio_users_use_address = true
     config.folio_users_omniauth_providers = %i[facebook google_oauth2 twitter]
     config.folio_users_after_ajax_sign_up_redirect = false